resolve React warning 8/8: extract shared transaction category select to remove duplicated category state logic
